pr review

+++ src/overlays/actors/ovl_Bg_Open_Shutter/z_bg_open_shutter.h
@@ -3,7 +3,7 @@
 
 #include "global.h"
 
-#define BGOPENSHUTTER_GET_A(thisx) ((thisx)->params) >> 0xA
+#define BGOPENSHUTTER_GET_A(thisx) ((u16)((thisx)->params) >> 0xA)
 
 struct BgOpenShutter;
